+++ packages/cli/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.51] - 2026-08-31
+
 ### Added
 
 - Add `--supabase-project-ref <ref>` for non-interactive, project-scoped